requirements.txt
streamlit
pandas
Pillow
openpyxl